Move tests to core/local and js packages
