--- js/modules/k6/execution/utils_test.go
@@ -1,109 +1,0 @@
-/*
- *
- * k6 - a next-generation load testing tool
- * Copyright (C) 2021 Load Impact
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package execution_test
-
-import (
-	"context"
-	"net/url"
-	"testing"
-
-	"github.com/sirupsen/logrus"
-	"github.com/spf13/afero"
-	"github.com/stretchr/testify/require"
-	"gopkg.in/guregu/null.v3"
-
-	"go.k6.io/k6/core/local"
-	"go.k6.io/k6/js"
-	"go.k6.io/k6/lib"
-	"go.k6.io/k6/lib/executor"
-	"go.k6.io/k6/lib/testutils"
-	"go.k6.io/k6/lib/testutils/minirunner"
-	"go.k6.io/k6/loader"
-	"go.k6.io/k6/stats"
-)
-
-// Copied from k6:js/console_test.go
-func getSimpleRunner(tb testing.TB, filename, data string, opts ...interface{}) (lib.Runner, error) {
-	var (
-		fs     = afero.NewMemMapFs()
-		rtOpts = lib.RuntimeOptions{CompatibilityMode: null.NewString("base", true)}
-		logger = testutils.NewLogger(tb)
-	)
-	for _, o := range opts {
-		switch opt := o.(type) {
-		case afero.Fs:
-			fs = opt
-		case lib.RuntimeOptions:
-			rtOpts = opt
-		case *logrus.Logger:
-			logger = opt
-		}
-	}
-	return js.New(
-		logger,
-		&loader.SourceData{
-			URL:  &url.URL{Path: filename, Scheme: "file"},
-			Data: []byte(data),
-		},
-		map[string]afero.Fs{"file": fs, "https": afero.NewMemMapFs()},
-		rtOpts,
-	)
-}
-
-// Copied from k6:core/local/local_test.go
-func newTestExecutionScheduler(
-	t *testing.T, runner lib.Runner, logger *logrus.Logger, opts lib.Options,
-) (ctx context.Context, cancel func(), execScheduler *local.ExecutionScheduler, samples chan stats.SampleContainer) {
-	if runner == nil {
-		runner = &minirunner.MiniRunner{}
-	}
-	ctx, cancel = context.WithCancel(context.Background())
-	newOpts, err := executor.DeriveScenariosFromShortcuts(lib.Options{
-		MetricSamplesBufferSize: null.NewInt(200, false),
-	}.Apply(runner.GetOptions()).Apply(opts))
-	require.NoError(t, err)
-	require.Empty(t, newOpts.Validate())
-
-	require.NoError(t, runner.SetOptions(newOpts))
-
-	if logger == nil {
-		logger = logrus.New()
-		logger.SetOutput(testutils.NewTestOutput(t))
-	}
-
-	execScheduler, err = local.NewExecutionScheduler(runner, logger)
-	require.NoError(t, err)
-
-	samples = make(chan stats.SampleContainer, newOpts.MetricSamplesBufferSize.Int64)
-	go func() {
-		for {
-			select {
-			case <-samples:
-			case <-ctx.Done():
-				return
-			}
-		}
-	}()
-
-	require.NoError(t, execScheduler.Init(ctx, samples))
-
-	return ctx, cancel, execScheduler, samples
-}
